Begin cleaning up some type mismatches, unused variables, and a deprecated function.

+++ Source/WebCore/bridge/qt/qt_runtime.cpp
@@ -263,7 +263,7 @@ QList<ItemType> convertToList(JSContextRef context, JSRealType type, JSObjectRef
             else
                 break;
         }
-        if (list.count() != length)
+        if (static_cast<unsigned>(list.count()) != length)
             list.clear();
         else if (distance)
             *distance = 5;
@@ -976,7 +976,7 @@ static int indexOfMetaEnum(const QMetaObject *meta, const QByteArray &str)
 static int findMethodIndex(JSContextRef context,
                            const QMetaObject* meta,
                            const QByteArray& signature,
-                           int argumentCount,
+                           unsigned int argumentCount,
                            const JSValueRef arguments[],
                            bool allowPrivate,
                            QVarLengthArray<QVariant, 10> &vars,
